requirements.txt
asgiref==3.9.1
Django==4.2.7
djangorestframework==3.14.0
mssql-django==1.3
Pillow==10.1.0
pymssql==2.3.7
pyodbc==5.0.1
python-dotenv==1.0.0
pytz==2025.2
sqlparse==0.5.3
